examples/news-reader/pyproject.toml
[project]
name = "news-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Paulescu", email = "plabartabajo@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "liquid-audio>=1.0.0",
    "modal>=1.2.4",
    "numpy>=2.3.5",
    "pyaudio>=0.2.14",
    "pygame>=2.6.0",
    "rich>=13.0.0",
    "scipy>=1.16.3",
    "torchcodec>=0.8.1",
]

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = false
force-sort-within-sections = true
known-first-party = ["news_reader"]
